pyproject.toml
[tool.poetry]
name = "api-facturas"
version = "1.0.0"
description = "FastAPI invoice extraction from WhatsApp ZIP files using Gemini AI"
authors = ["Elvis Cruz <elvis@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
python-multipart = "^0.0.6"
google-cloud-aiplatform = "^1.38.1"
PyPDF2 = "^3.0.1"
pdfplumber = "^0.10.3"
Pillow = "^10.1.0"
python-magic = "^0.4.27"
chardet = "^5.2.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
PyMuPDF = "^1.23.14"
vertexai = "^1.38.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[tool.poetry.scripts]
dev = "main:run_dev"
start = "main:run_dev"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

